Update single furniture page
