+++ src/App.js
@@ -2,7 +2,8 @@ import { Route, Routes } from 'react-router-dom';
 import './App.css';
 import Login from './components/login/login';
 import NotFound from './components/pagenotfound.js/notfound';
-import Furnitures from './components/dashboard/furniture/Furnitures';
+import FurnitureList from './components/dashboard/furniture/FurnitureList';
+import Furniture from './components/dashboard/furniture/Furniture';
 import Customer from './components/dashboard/customer/customer';
 import Product from './components/dashboard/product/product';
 import Reserve from './components/dashboard/reserve/reserve';
@@ -14,8 +15,9 @@ function App() {
   return (
     <Routes>
       <Route path="/" exact element={<Home />}>
-        <Route path="/" exact element={<Furnitures />} />
-        <Route path="/furnitures" exact element={<Furnitures />} />
+        <Route path="/" exact element={<FurnitureList />} />
+        <Route path="/furnitures" exact element={<FurnitureList />} />
+        <Route path="/furnitures/:id" element={<Furniture />} />
         <Route path="/customers" exact element={<Customer />} />
         <Route path="/products" exact element={<Product />} />
         <Route path="/reserves" exact element={<Reserve />} />
